crypto_trading_agency/requirements.txt
agency-swarm>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.1.0
ta>=0.10.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
plotly>=5.18.0
requests>=2.31.0
python-binance>=1.0.19
websocket-client>=1.7.0
aiohttp>=3.9.0
joblib>=1.3.0
websockets>=12.0
torch>=2.1.0
xgboost>=2.0.0
lightgbm>=4.1.0
statsmodels>=0.14.0
scipy>=1.11.0
pandas-ta>=0.3.14b 